Rework nvim config

+++ home/g/nvim/lua/core/pack.lua
@@ -1,9 +1,9 @@
-local fn,uv,api = vim.fn,vim.loop,vim.api
+local fn, uv, api = vim.fn, vim.loop, vim.api
 local vim_path = require('core.global').vim_path
 local data_dir = require('core.global').data_dir
 local modules_dir = vim_path .. '/lua/modules'
-local packer_compiled = data_dir..'packer_compiled.vim'
-local compile_to_lua = data_dir..'lua/_compiled.lua'
+local packer_compiled = data_dir .. 'packer_compiled.vim'
+local compile_to_lua = data_dir .. 'lua/_compiled.lua'
 local packer = nil
 
 local Packer = {}
@@ -12,20 +12,18 @@ Packer.__index = Packer
 function Packer:load_plugins()
   self.repos = {}
 
-  local get_plugins_list = function ()
+  local get_plugins_list = function()
     local list = {}
-    local tmp = vim.split(fn.globpath(modules_dir,'*/plugins.lua'),'\n')
-    for _,f in ipairs(tmp) do
-      list[#list+1] = f:sub(#modules_dir - 6,-1)
-    end
+    local tmp = vim.split(fn.globpath(modules_dir, '*/plugins.lua'), '\n')
+    for _, f in ipairs(tmp) do list[#list + 1] = f:sub(#modules_dir - 6, -1) end
     return list
   end
 
   local plugins_file = get_plugins_list()
-  for _,m in ipairs(plugins_file) do
-    local repos = require(m:sub(0,#m-4))
-    for repo,conf in pairs(repos) do
-      self.repos[#self.repos+1] = vim.tbl_extend('force',{repo},conf)
+  for _, m in ipairs(plugins_file) do
+    local repos = require(m:sub(0, #m - 4))
+    for repo, conf in pairs(repos) do
+      self.repos[#self.repos + 1] = vim.tbl_extend('force', {repo}, conf)
     end
   end
 end
@@ -37,27 +35,25 @@ function Packer:load_packer()
   end
   packer.init({
     compile_path = packer_compiled,
-    git = { clone_timeout = 120 },
-    disable_commands = true
+    git = {clone_timeout = 120},
+    disable_commands = true,
   })
   packer.reset()
   local use = packer.use
   self:load_plugins()
-  use {"wbthomason/packer.nvim", opt = true }
-  for _,repo in ipairs(self.repos) do
-    use(repo)
-  end
+  use {"wbthomason/packer.nvim", opt = true}
+  for _, repo in ipairs(self.repos) do use(repo) end
 end
 
 function Packer:init_ensure_plugins()
-  local packer_dir = data_dir..'pack/packer/opt/packer.nvim'
+  local packer_dir = data_dir .. 'pack/packer/opt/packer.nvim'
   local state = uv.fs_stat(packer_dir)
   if not state then
-    local cmd = "!git clone https://github.com/wbthomason/packer.nvim " ..packer_dir
+    local cmd = "!git clone https://github.com/wbthomason/packer.nvim " ..
+                  packer_dir
     api.nvim_command(cmd)
-    uv.fs_mkdir(data_dir..'lua',511,function()
-      assert("make compile path dir faield")
-    end)
+    uv.fs_mkdir(data_dir .. 'lua', 511,
+                function() assert("make compile path dir faield") end)
     self:load_packer()
     packer.install()
   end
@@ -65,45 +61,47 @@ end
 
 local plugins = setmetatable({}, {
   __index = function(_, key)
-    if not packer then
-      Packer:load_packer()
-    end
+    if not packer then Packer:load_packer() end
     return packer[key]
-  end
+  end,
 })
 
-function plugins.ensure_plugins()
-  Packer:init_ensure_plugins()
+function plugins.ensure_plugins() Packer:init_ensure_plugins() end
+
+function file_exists(name)
+  local f = io.open(name, "r")
+  if f ~= nil then
+    io.close(f)
+    return true
+  else
+    return false
+  end
 end
 
 function plugins.convert_compile_file()
   local lines = {}
   local lnum = 1
-  lines[#lines+1] = 'vim.cmd [[packadd packer.nvim]]\n'
+  lines[#lines + 1] = 'vim.cmd [[packadd packer.nvim]]\n'
 
-  for line in io.lines(packer_compiled) do
-    lnum = lnum + 1
-    if lnum > 15 then
-      lines[#lines+1] = line .. '\n'
-      if line == 'END' then
-        break
+  if file_exists(packer_compiled) then
+    for line in io.lines(packer_compiled) do
+      lnum = lnum + 1
+      if lnum > 15 then
+        lines[#lines + 1] = line .. '\n'
+        if line == 'END' then break end
       end
     end
   end
-  table.remove(lines,#lines)
+  table.remove(lines, #lines)
 
   if vim.fn.isdirectory(data_dir .. 'lua') ~= 1 then
-    os.execute('mkdir -p '..data_dir .. 'lua')
+    os.execute('mkdir -p ' .. data_dir .. 'lua')
   end
 
-  if vim.fn.filereadable(compile_to_lua) == 1 then
-    os.remove(compile_to_lua)
-  end
+  if vim.fn.filereadable(compile_to_lua) == 1 then os.remove(compile_to_lua) end
 
-  local file = io.open(compile_to_lua,"w")
-  for _,line in ipairs(lines) do
-    file:write(line)
-  end
+  local file = io.open(compile_to_lua, "w")
+  for _, line in ipairs(lines) do file:write(line) end
   file:close()
 
   os.remove(packer_compiled)
@@ -127,7 +125,8 @@ function plugins.load_compile()
   if vim.fn.filereadable(compile_to_lua) == 1 then
     require('_compiled')
   else
-    assert('Missing packer compile file Run PackerCompile Or PackerInstall to fix')
+    assert(
+      'Missing packer compile file Run PackerCompile Or PackerInstall to fix')
   end
   vim.cmd [[command! PackerCompile lua require('core.pack').magic_compile()]]
   vim.cmd [[command! PackerInstall lua require('core.pack').install()]]
